refactor(openomni): split resident runtime responsibilities
